pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "databaseparsers"
version = "1.0"
description = "Collection of NOMAD parsers for databases."
readme = "README.md"
authors = [{ name = "The NOMAD Authors" }]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.6.dev1",
    "nomad-schema-plugin-simulation-workflow@git+https://github.com/nomad-coe/nomad-schema-plugin-simulation-workflow.git@develop",
    "nomad-schema-plugin-run@git+https://github.com/nomad-coe/nomad-schema-plugin-run.git@develop",
    "pydantic>=1.10.8,<2.0.0",
]

[project.urls]
homepage = "https://github.com/nomad-coe/database-parsers"

[project.optional-dependencies]
tests = [
    "mypy==1.0.1",
    "pylint==2.13.9",
    "pylint_plugin_utils==0.5",
    "pycodestyle==2.8.0",
    "pytest==3.10.0",
    "pytest-timeout==1.4.2",
    "pytest-cov==2.7.1",
    "astroid==2.11.7",
    "typing-extensions==4.4.0",
    "ruff==0.1.4",
]

[tool.ruff]
include = ["electronicparsers/*.py", "tests/*.py"]
select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "PL", # pylint
]
ignore = [
    "E501", # Line too long ({width} > {limit} characters)
    "E701", # Multiple statements on one line (colon)
    "E731", # Do not assign a lambda expression, use a def
    "E402",  # Module level import not at top of file
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used instead of constant
    "PLW0603", # Using the global statement
    "PLW2901", # redefined-loop-name
    "PLR1714", # consider-using-in
    "PLR5501", # else-if-used
]
fixable = ["ALL"]

[tool.setuptools.packages.find]
include = [
    "databaseparsers*",
]

[project.entry-points.'nomad.plugin']
openkimparser = "databaseparsers:openkim_parser_entry_point"

[tool.mypy]
strict = false
ignore_missing_imports = true
follow_imports = "silent"
no_strict_optional = true
disable_error_code = "import, annotation-unchecked"
